feat: ok

+++ src/api/player/delete.js
@@ -5,10 +5,19 @@ import model from "../../db/model/index";
 
 export default async (req, res) => {
   const { id } = req.params;
+  // check user exists
+  const existsPlayer = await model.Player.findOne({
+    where: { id },
+  });
+  if (!existsPlayer) {
+    res.status(404).send("Player not found");
+    return;
+  }
+
   const deletedPlayer = await model.Player.destroy({
     where: { id },
     include: [{ model: model.PlayerSkill }],
   });
 
-  res.send(deletedPlayer).status(200);
-}
+  res.status(200).send(deletedPlayer);
+};
